@@ -1,55 +1,52 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20141013223344) do
+ActiveRecord::Schema.define(version: 2014_10_13_223344) do
 
-  create_table "categories", force: true do |t|
-    t.string   "name"
+  create_table "categories", force: :cascade do |t|
+    t.string "name"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "credit_items", force: true do |t|
-    t.integer  "user_id"
-    t.integer  "customer_id"
-    t.integer  "category_id",                         null: false
-    t.decimal  "value",       precision: 8, scale: 2, null: false
-    t.text     "note"
+  create_table "credit_items", force: :cascade do |t|
+    t.integer "user_id"
+    t.integer "customer_id"
+    t.integer "category_id", null: false
+    t.decimal "value", precision: 8, scale: 2, null: false
+    t.text "note"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "customers", force: true do |t|
-    t.string   "first_name",  null: false
-    t.string   "last_name",   null: false
-    t.string   "email"
-    t.string   "postal_code"
-    t.string   "customer_id"
-    t.string   "view_token",  null: false
+  create_table "customers", force: :cascade do |t|
+    t.string "first_name", null: false
+    t.string "last_name", null: false
+    t.string "email"
+    t.string "postal_code"
+    t.string "customer_id"
+    t.string "view_token", null: false
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.index ["customer_id"], name: "index_customers_on_customer_id", unique: true
+    t.index ["first_name"], name: "index_customers_on_first_name"
+    t.index ["last_name"], name: "index_customers_on_last_name"
   end
 
-  add_index "customers", ["customer_id"], name: "index_customers_on_customer_id", unique: true
-  add_index "customers", ["first_name"], name: "index_customers_on_first_name"
-  add_index "customers", ["last_name"], name: "index_customers_on_last_name"
-
-  create_table "users", force: true do |t|
-    t.string "email",                            null: false
-    t.string "password_digest",                  null: false
-    t.string "role",            default: "user", null: false
+  create_table "users", force: :cascade do |t|
+    t.string "email", null: false
+    t.string "password_digest", null: false
+    t.string "role", default: "user", null: false
+    t.index ["email"], name: "index_users_on_email", unique: true
   end
-
-  add_index "users", ["email"], name: "index_users_on_email", unique: true
 
 end
